Updated Angular to latest versions (10.x).
As a result of the update, the 'haversine-position' library no longer
worked.  It gave a 'ReferenceError: exports is not defined' error in the
browser console.  It appears that the reason for this is that the
library is written in Typescript, and then compiled to JavaScript for
distribution, and the configuration of the Typescript compiler in the
module is not compatible with the newer Angular settings, which
discourage 'commonjs' modules.

The problem was fixed by incorporating the 'haversine-position' library
source code into the local-aprs source tree (the module is MIT licensed,
so that's OK).